refactor(ui): extract composer mention hook + workspace row (#1044)
Move the @// mention-popup state machine (trigger detection, items,
active index, post-insertion suppression, selectionchange listener) into
useMentionPopup; Composer keeps only the keydown routing, driving the
hook through identically-named bindings so the mention contract reads
the same source shape.

Move the workspace + branch picker row into ComposerWorkspaceRow with
named picker prop types; project-context-badge is co-migrated to pin the
same invariants across the composer/row seam, and the copy-hygiene +
stub-vocabulary scans follow the moved file. Also drops two imports that
became dead with the move (MenuSeparator, FileEdit).
